+++ packages/7715-permission-types/test/permissions/rules/expiry.test.ts
@@ -1,18 +1,14 @@
 import { createTimestampTerms } from '@metamask/delegation-core';
-import {
-  CHAIN_ID,
-  DELEGATOR_CONTRACTS,
-} from '@metamask/delegation-deployments';
 import type { Hex } from '@metamask/utils';
 import { describe, it, expect } from 'vitest';
 
 import { expiryRuleDecoder } from '../../../src/permissions/rules/expiry';
 import type { ChecksumCaveat } from '../../../src/permissions/types';
-import { getChecksumEnforcersByChainId } from '../../../src/permissions/utils';
+import { checksumEnforcerAddresses } from '../../../src/permissions/utils';
+import { contracts } from '../../test-utils';
 
 describe('expiryRule', () => {
-  const contracts = DELEGATOR_CONTRACTS['1.3.0'][CHAIN_ID.sepolia];
-  const contractAddresses = getChecksumEnforcersByChainId(contracts);
+  const contractAddresses = checksumEnforcerAddresses(contracts);
   const { timestampEnforcer, nonceEnforcer } = contractAddresses;
   const requiredEnforcers = new Map<Hex, number>([[nonceEnforcer, 1]]);